+++ src/websocket/server.cc
@@ -25,22 +25,16 @@
 #include <seastar/util/defer.hh>
 #include <seastar/util/log.hh>
 #include <seastar/core/scattered_message.hh>
-#include <seastar/core/byteorder.hh>
 #include <seastar/http/request.hh>
-#include <gnutls/crypto.h>
-#include <gnutls/gnutls.h>
 
 namespace seastar::experimental::websocket {
 
-static sstring magic_key_suffix = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";
 static sstring http_upgrade_reply_template =
     "HTTP/1.1 101 Switching Protocols\r\n"
     "Upgrade: websocket\r\n"
     "Connection: Upgrade\r\n"
     "Sec-WebSocket-Version: 13\r\n"
     "Sec-WebSocket-Accept: ";
-
-static logger wlogger("websocket");
 
 void server::listen(socket_address addr, listen_options lo) {
     _listeners.push_back(seastar::listen(addr, lo));
@@ -112,27 +106,6 @@ future<> server_connection::process() {
     });
 }
 
-static std::string sha1_base64(std::string_view source) {
-    unsigned char hash[20];
-    assert(sizeof(hash) == gnutls_hash_get_len(GNUTLS_DIG_SHA1));
-    if (int ret = gnutls_hash_fast(GNUTLS_DIG_SHA1, source.data(), source.size(), hash);
-        ret != GNUTLS_E_SUCCESS) {
-        throw websocket::exception(fmt::format("gnutls_hash_fast: {}", gnutls_strerror(ret)));
-    }
-    gnutls_datum_t hash_data{
-        .data = hash,
-        .size = sizeof(hash),
-    };
-    gnutls_datum_t base64_encoded;
-    if (int ret = gnutls_base64_encode2(&hash_data, &base64_encoded);
-        ret != GNUTLS_E_SUCCESS) {
-        throw websocket::exception(fmt::format("gnutls_base64_encode2: {}", gnutls_strerror(ret)));
-    }
-    auto free_base64_encoded = defer([&] () noexcept { gnutls_free(base64_encoded.data); });
-    // base64_encoded.data is "unsigned char *"
-    return std::string(reinterpret_cast<const char*>(base64_encoded.data), base64_encoded.size);
-}
-
 future<> server_connection::read_http_upgrade_request() {
     _http_parser.init();
     co_await _read_buf.consume(_http_parser);
@@ -180,51 +153,6 @@ future<> server_connection::read_http_upgrade_request() {
     co_await _write_buf.flush();
 }
 
-future<> server_connection::handle_ping() {
-    // TODO
-    return make_ready_future<>();
-}
-
-future<> server_connection::handle_pong() {
-    // TODO
-    return make_ready_future<>();
-}
-
-
-future<> server_connection::read_one() {
-    return _read_buf.consume(_websocket_parser).then([this] () mutable {
-        if (_websocket_parser.is_valid()) {
-            // FIXME: implement error handling
-            switch(_websocket_parser.opcode()) {
-                // We do not distinguish between these 3 types.
-                case opcodes::CONTINUATION:
-                case opcodes::TEXT:
-                case opcodes::BINARY:
-                    return _input_buffer.push_eventually(_websocket_parser.result());
-                case opcodes::CLOSE:
-                    wlogger.debug("Received close frame.");
-                    /*
-                     * datatracker.ietf.org/doc/html/rfc6455#section-5.5.1
-                     */
-                    return close(true);
-                case opcodes::PING:
-                    wlogger.debug("Received ping frame.");
-                    return handle_ping();
-                case opcodes::PONG:
-                    wlogger.debug("Received pong frame.");
-                    return handle_pong();
-                default:
-                    // Invalid - do nothing.
-                    ;
-            }
-        } else if (_websocket_parser.eof()) {
-            return close(false);
-        }
-        wlogger.debug("Reading from socket has failed.");
-        return close(true);
-    });
-}
-
 future<> server_connection::read_loop() {
     return read_http_upgrade_request().then([this] {
         return when_all_succeed(
@@ -237,63 +165,6 @@ future<> server_connection::read_loop() {
         ).discard_result();
     }).finally([this] {
         return _read_buf.close();
-    });
-}
-
-void server_connection::shutdown_input() {
-    _fd.shutdown_input();
-}
-
-future<> server_connection::close(bool send_close) {
-    return [this, send_close]() {
-        if (send_close) {
-            return send_data(opcodes::CLOSE, temporary_buffer<char>(0));
-        } else {
-            return make_ready_future<>();
-        }
-    }().finally([this] {
-        _done = true;
-        return when_all_succeed(_input.close(), _output.close()).discard_result().finally([this] {
-            _fd.shutdown_output();
-        });
-    });
-}
-
-future<> server_connection::send_data(opcodes opcode, temporary_buffer<char>&& buff) {
-    char header[10] = {'\x80', 0};
-    size_t header_size = 2;
-
-    header[0] += opcode;
-
-    if ((126 <= buff.size()) && (buff.size() <= std::numeric_limits<uint16_t>::max())) {
-        header[1] = 0x7E;
-        write_be<uint16_t>(header + 2, buff.size());
-        header_size += sizeof(uint16_t);
-    } else if (std::numeric_limits<uint16_t>::max() < buff.size()) {
-        header[1] = 0x7F;
-        write_be<uint64_t>(header + 2, buff.size());
-        header_size += sizeof(uint64_t);
-    } else {
-        header[1] = uint8_t(buff.size());
-    }
-
-    scattered_message<char> msg;
-    msg.append(sstring(header, header_size));
-    msg.append(std::move(buff));
-    return _write_buf.write(std::move(msg)).then([this] {
-        return _write_buf.flush();
-    });
-}
-
-future<> server_connection::response_loop() {
-    return do_until([this] {return _done;}, [this] {
-        // FIXME: implement error handling
-        return _output_buffer.pop_eventually().then([this] (
-                temporary_buffer<char> buf) {
-            return send_data(opcodes::BINARY, std::move(buf));
-        });
-    }).finally([this]() {
-        return _write_buf.close();
     });
 }
